Update yt-dlpan.sh
Added a feature to findremoved. Prevents the function from running if it was last run on that folder very recently.
This is necessary since this is features makes too many requests and consumes a lot of time. I had tested the statement and ensured it works, but I can't use a real scenario as a test right now.

+++ yt-dlpan.sh
@@ -32,31 +32,36 @@ antiban='--sleep-requests 0.5 --min-sleep-interval 3 --max-sleep-interval 20' # 
 
 # These functions will run after you finish downloading all the files in a parent directory.
 function findremoved() {
-	echo "Detecting deleted videos, to link them to another folder"
-	# rate limits won't break this. You don't need cookies. $antiban is optional if you are concerned about ip ban.
-	local parent="$1"
-	local target="$2"
-	local archive="$3"
-	local tracking="$4"
-	rm "${idlists}"/offline.txt; rm "${idlists}"/found.txt
-	if [ $tracking ]; then
- 		# Antiban is not necessary since the error won't interfere. It's only here to play it safe. I suggest you take the risk and remove it since it will take so much longer with antiban.
-		yt-dlp ${antiban} -s ${target} 2>&1 >/dev/null | perl -pe "s/(${logscleaner})//g" | sed -r '/^\s*$/d' | tee offline.txt # detector of deleted videos. To get full logs, which also serves as a progress bar, replace "2>&1 >/dev/null" with "2> >(tee >(cat 1>&2) pipes)". Full logs are the same as download, so they are limited by default. Warbo stackoverflow 45798436
-		found=$(comm -1 -2 <(sort "${archive}") <(sort offline.txt)); echo "$found" > found.txt # Gets only the deleted files that had been downloaded previously. This is unnecessary, and it only reduces the amount of commands, the log size(arguably), and it insignificantly decreases time.
-	else
- 		# Antiban is not necessary because it's only a few requests.
-		yt-dlp --download-archive offline.txt --force-write-archive --flat-playlist -s ${target} >/dev/null 2>&1 # lists every video the channel has. To get full logs, which also serves as a progress bar, remove >/dev/null 2>&1. There's no logs by default because this is similar to your download command.
-		found=$(comm -2 -3 <(sort "${archive}") <(sort offline.txt)); echo "$found" > found.txt # if a video is not on the channel, but you have it downloaded, it's assumed that it has been removed.
-	fi
-	sed -i -r "s/(${websites})//g" found.txt
-	for gone in $(cat found.txt); do if ! grep -Exq "${parent}/preserving/.* \[$gone\]\..*" "${idlists}"/foundremoved.txt; then # If user deletes a file from folder, it won't be recopied. This is optional, feel free to remove or disable.
-		if ( [ -f "${parent}"/*$gone* ] || [ -f "${parent}"/thumbs/*$gone* ] ) && [[ $direxists != true ]]; then mkdir -p "${parent}"/preserving/thumbs; local direxists=true; fi # creates folder only if you have a removed file. This is a good notification, and the if is necessary to prevent spam.
-		cp -vs "${parent}"/*"$gone"* "${parent}"/preserving/
-		cp -vs "${parent}"/thumbs/*"$gone"* "${parent}"/preserving/thumbs
-		find "${parent}"/preserving/ -name "*$gone*" >> "${idlists}"/foundremoved.txt 
-	fi; done
-	mv -vf "${idlists}"/offline.txt "${parent}"/preserving/ # full list
-	mv -vf "${idlists}"/found.txt "${parent}"/preserving/ # what you are preserving
+	if [ $(( $(date +%s) - ($(date +%s -r "${parent}"/preserving/found.txt)+0) )) -le 7889238 ]; then # Currently set to skip if it was last run less than 3 months old.
+		# Necessary since this makes too many requests and takes too much time. This can only be calculated in seconds if you want to change this.
+  		echo "Not checking for deleted videos because the last check was too recent"
+    	else
+		echo "Detecting deleted videos, to link them to another folder"
+		# rate limits won't break this. You don't need cookies. $antiban is optional if you are concerned about ip ban.
+		local parent="$1"
+		local target="$2"
+		local archive="$3"
+		local tracking="$4"
+		rm "${idlists}"/offline.txt; rm "${idlists}"/found.txt
+		if [ $tracking ]; then
+	 		# Antiban is not necessary since the error won't interfere. It's only here to play it safe. I suggest you take the risk and remove it since it will take so much longer with antiban.
+			yt-dlp ${antiban} -s ${target} 2>&1 >/dev/null | perl -pe "s/(${logscleaner})//g" | sed -r '/^\s*$/d' | tee offline.txt # detector of deleted videos. To get full logs, which also serves as a progress bar, replace "2>&1 >/dev/null" with "2> >(tee >(cat 1>&2) pipes)". Full logs are the same as download, so they are limited by default. Warbo stackoverflow 45798436
+			found=$(comm -1 -2 <(sort "${archive}") <(sort offline.txt)); echo "$found" > found.txt # Gets only the deleted files that had been downloaded previously. This is unnecessary, and it only reduces the amount of commands, the log size(arguably), and it insignificantly decreases time.
+		else
+	 		# Antiban is not necessary because it's only a few requests.
+			yt-dlp --download-archive offline.txt --force-write-archive --flat-playlist -s ${target} >/dev/null 2>&1 # lists every video the channel has. To get full logs, which also serves as a progress bar, remove >/dev/null 2>&1. There's no logs by default because this is similar to your download command.
+			found=$(comm -2 -3 <(sort "${archive}") <(sort offline.txt)); echo "$found" > found.txt # if a video is not on the channel, but you have it downloaded, it's assumed that it has been removed.
+		fi
+		sed -i -r "s/(${websites})//g" found.txt
+		for gone in $(cat found.txt); do if ! grep -Exq "${parent}/preserving/.* \[$gone\]\..*" "${idlists}"/foundremoved.txt; then # If user deletes a file from folder, it won't be recopied. This is optional, feel free to remove or disable.
+			if ( [ -f "${parent}"/*$gone* ] || [ -f "${parent}"/thumbs/*$gone* ] ) && [[ $direxists != true ]]; then mkdir -p "${parent}"/preserving/thumbs; local direxists=true; fi # creates folder only if you have a removed file. This is a good notification, and the if is necessary to prevent spam.
+			cp -vs "${parent}"/*"$gone"* "${parent}"/preserving/
+			cp -vs "${parent}"/thumbs/*"$gone"* "${parent}"/preserving/thumbs
+			find "${parent}"/preserving/ -name "*$gone*" >> "${idlists}"/foundremoved.txt 
+		fi; done
+		mv -vf "${idlists}"/offline.txt "${parent}"/preserving/ # full list
+		mv -vf "${idlists}"/found.txt "${parent}"/preserving/ # what you are preserving
+  	fi
 }
 function frugalizer() { # provides a video of much lower filesize than remnant.
 	local parent="$1"
